perf: reduce uneeded allocs in solver
