+++ constraint/blueprint.go
@@ -1,16 +1,28 @@
 package constraint
 
+type BlueprintID uint32
+
+// Blueprint enable representing heterogenous constraints or instructions in a constraint system
+// in a memory efficient way. Blueprints essentially help the frontend/ to "compress"
+// constraints or instructions, and specify for the solving (or zksnark setup) part how to
+// "decompress" and optionally "solve" the associated wires.
 type Blueprint interface {
 	// NbInputs return the number of calldata input this blueprint expects.
 	// If this is unknown at compile time, implementation must return -1 and store
 	// the actual number of inputs in the first index of the calldata.
 	NbInputs() int
 
-	// NbConstraints return the number of constraints this blueprint creates. For a hint, that's 0.
+	// NbConstraints return the number of constraints this blueprint creates.
 	NbConstraints() int
+}
 
-	// // Wires fills the wires that appear in the instantiation of this blueprint.
-	// Wires(calldata []uint32, wires *[]uint32)
+// Solver represents the state of a constraint system solver at runtime. Blueprint can interact
+// with this object to perform run time logic, solve constraints and assign values in the solution.
+type Solver interface {
+	Field
+	GetValue(cID, vID uint32) Element
+	GetCoeff(cID uint32) Element
+	SetValue(vID uint32, f Element)
 }
 
 // BlueprintSolvable represents a blueprint that knows how to solve itself.
@@ -35,9 +47,4 @@ type BlueprintSparseR1C interface {
 type BlueprintHint interface {
 	CompressHint(HintMapping) []uint32
 	DecompressHint(h *HintMapping, calldata []uint32)
-}
-
-type BlueprintSparseR1CBlock interface {
-	CompressBlock()
-	DecompressBlock() []SparseR1C
 }
