+++ internal/generator/backend/template/representations/solver.go.tmpl
@@ -167,7 +167,7 @@ func (s *solver) accumulateInto(t constraint.Term, r *fr.Element) {
 }
 
 // solveWithHint executes a hint and assign the result to its defined outputs.
-func (s *solver) solveWithHint(h constraint.HintMapping) error {
+func (s *solver) solveWithHint(h *constraint.HintMapping) error {
 	// ensure hint function was provided
 	f, ok := s.mHintsFunctions[h.HintID]
 	if !ok {
@@ -368,9 +368,8 @@ func (solver *solver) processInstruction(inst constraint.Instruction, scratch *s
 	// blueprint encodes a hint, we execute.
 	// TODO @gbotrel may be worth it to move hint logic in blueprint "solve"
 	if bc, ok := blueprint.(constraint.BlueprintHint); ok {
-		var hm constraint.HintMapping
-		bc.DecompressHint(&hm, calldata)
-		return solver.solveWithHint(hm)
+		bc.DecompressHint(&scratch.tHint, calldata)
+		return solver.solveWithHint(&scratch.tHint)
 	}
 
 
@@ -752,5 +751,6 @@ func (solver *solver) wrapErrWithDebugInfo(cID uint32, err error) *UnsatisfiedCo
 type scratch struct {
 	tR1C constraint.R1C
 	tSparseR1C constraint.SparseR1C
+	tHint constraint.HintMapping
 }
